refactor(agentic-ai): unify OAuth 2.0 client-credentials config across providers
OpenAI and Anthropic custom-backend authentication both target a plain
OAuth 2.0 authorization server the same way, unlike apiKey, which is
provider-specific. Extract the previously-duplicated
OAuthClientCredentialsAuthentication record into a single shared type
implementing both OpenAiCustomEndpointAuthentication and
AnthropicCustomEndpointAuthentication, so future OAuth 2.0 config
additions in the AI agent module don't repeat this shape.
