+++ connectors/agentic-ai/connector-agentic-ai/src/main/java/io/camunda/connector/agenticai/aiagent/model/request/v2/AnthropicCustomEndpointAuthentication.java
@@ -8,19 +8,11 @@ package io.camunda.connector.agenticai.aiagent.model.request.v2;
 
 import com.fasterxml.jackson.annotation.JsonSubTypes;
 import com.fasterxml.jackson.annotation.JsonTypeInfo;
-import io.camunda.connector.api.annotation.FEEL;
 import io.camunda.connector.generator.java.annotation.FeelMode;
 import io.camunda.connector.generator.java.annotation.TemplateDiscriminatorProperty;
 import io.camunda.connector.generator.java.annotation.TemplateProperty;
-import io.camunda.connector.generator.java.annotation.TemplateProperty.DropdownPropertyChoice;
-import io.camunda.connector.generator.java.annotation.TemplateProperty.PropertyType;
 import io.camunda.connector.generator.java.annotation.TemplateSubType;
-import io.camunda.connector.http.client.authentication.OAuthConstants;
 import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.NotEmpty;
-import jakarta.validation.constraints.NotNull;
-import jakarta.validation.constraints.Pattern;
-import org.apache.commons.lang3.builder.ToStringBuilder;
 
 /**
  * Authentication strategies for Anthropic's {@code custom}-backend endpoint. Anthropic-compatible
@@ -37,8 +29,8 @@ import org.apache.commons.lang3.builder.ToStringBuilder;
       value = AnthropicCustomEndpointAuthentication.ApiKeyAuthentication.class,
       name = "apiKey"),
   @JsonSubTypes.Type(
-      value = AnthropicCustomEndpointAuthentication.OAuthClientCredentialsAuthentication.class,
-      name = AnthropicCustomEndpointAuthentication.OAuthClientCredentialsAuthentication.TYPE)
+      value = OAuthClientCredentialsAuthentication.class,
+      name = OAuthClientCredentialsAuthentication.TYPE)
 })
 @TemplateDiscriminatorProperty(
     label = "Authentication",
@@ -46,7 +38,10 @@ import org.apache.commons.lang3.builder.ToStringBuilder;
     name = "type",
     defaultValue = "none",
     description = "Authentication for the compatible API.")
-public sealed interface AnthropicCustomEndpointAuthentication {
+public sealed interface AnthropicCustomEndpointAuthentication
+    permits AnthropicCustomEndpointAuthentication.NoAuthentication,
+        AnthropicCustomEndpointAuthentication.ApiKeyAuthentication,
+        OAuthClientCredentialsAuthentication {
 
   @TemplateSubType(id = "none", label = "None")
   record NoAuthentication() implements AnthropicCustomEndpointAuthentication {}
@@ -66,102 +61,6 @@ public sealed interface AnthropicCustomEndpointAuthentication {
     @Override
     public String toString() {
       return "ApiKeyAuthentication{apiKey=[REDACTED]}";
-    }
-  }
-
-  @TemplateSubType(
-      id = AnthropicCustomEndpointAuthentication.OAuthClientCredentialsAuthentication.TYPE,
-      label = "OAuth 2.0")
-  record OAuthClientCredentialsAuthentication(
-      @FEEL
-          @NotEmpty
-          @Pattern(
-              regexp = "^(=|(http://|https://|secrets|\\{\\{).*$)",
-              message = "Must be a http(s) URL")
-          @TemplateProperty(
-              group = "provider",
-              tooltip = "The OAuth token endpoint",
-              label = "OAuth 2.0 token endpoint")
-          String oauthTokenEndpoint,
-      @FEEL
-          @NotEmpty
-          @TemplateProperty(
-              group = "provider",
-              tooltip = "Your application's client ID from the OAuth client",
-              label = "Client ID")
-          String clientId,
-      @FEEL
-          @NotEmpty
-          @TemplateProperty(
-              group = "provider",
-              tooltip = "Your application's client secret from the OAuth client",
-              label = "Client secret")
-          String clientSecret,
-      @FEEL
-          @TemplateProperty(
-              group = "provider",
-              tooltip = "The unique identifier of the target API you want to access",
-              optional = true)
-          String audience,
-      @NotNull
-          @TemplateProperty(
-              group = "provider",
-              type = PropertyType.Dropdown,
-              choices = {
-                @DropdownPropertyChoice(
-                    value = "BASIC_AUTH_HEADER",
-                    label = "Send as Basic Auth header"),
-                @DropdownPropertyChoice(
-                    value = "CREDENTIALS_BODY",
-                    label = "Send client credentials in body")
-              },
-              defaultValue = "BASIC_AUTH_HEADER",
-              tooltip =
-                  "Send client ID and client secret as Basic Auth request in the header, or as client credentials in the request body")
-          ClientAuthenticationMethod clientAuthentication,
-      @FEEL
-          @TemplateProperty(
-              group = "provider",
-              tooltip =
-                  "The scopes which you want to request authorization for (e.g.read:contacts)",
-              optional = true)
-          String scopes)
-      implements AnthropicCustomEndpointAuthentication {
-
-    @TemplateProperty(ignore = true)
-    public static final String TYPE = "oauth-client-credentials-flow";
-
-    public enum ClientAuthenticationMethod {
-      BASIC_AUTH_HEADER(OAuthConstants.BASIC_AUTH_HEADER),
-      CREDENTIALS_BODY(OAuthConstants.CREDENTIALS_BODY);
-
-      private final String oauthConstant;
-
-      ClientAuthenticationMethod(String oauthConstant) {
-        this.oauthConstant = oauthConstant;
-      }
-
-      public String oauthConstant() {
-        return oauthConstant;
-      }
-    }
-
-    public OAuthClientCredentialsAuthentication {
-      if (clientAuthentication == null) {
-        clientAuthentication = ClientAuthenticationMethod.BASIC_AUTH_HEADER;
-      }
-    }
-
-    @Override
-    public String toString() {
-      return new ToStringBuilder(this)
-          .append("oauthTokenEndpoint", oauthTokenEndpoint)
-          .append("clientId", "[REDACTED]")
-          .append("clientSecret", "[REDACTED]")
-          .append("audience", audience)
-          .append("clientAuthentication", clientAuthentication)
-          .append("scopes", scopes)
-          .toString();
     }
   }
 }
